Address requested changes
# Conflicts:
#	NitroxClient/GameLogic/PlayerLogic/PlayerCinematics.cs
#	NitroxPatcher/Patches/Dynamic/uGUI_SceneIntro_IntroSequence_Patch.cs
